persistence_app/requirements.txt
redis>=5,<6
python-dotenv>=1,<2
pymongo>=4.6,<5
